requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.4
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==3.4.8
dbus-python==1.2.18
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.47.0
netifaces==0.11.0
oauthlib==3.2.0
pycparser==2.21
pycryptodomex==3.19.0
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-decouple==3.8
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
systemd-python==234
urllib3==2.0.6
wadllib==1.3.6
websockets==11.0.3
yarl==1.9.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@00ef748cc0e35ee60efd0f7a00e373ab8d1af86b
youtube-search-python==1.6.6
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
zipp==1.0.0